Fixed Display
This is a VERY IMPORTANT commit. In this commit, 2 improvements were made.

The first improvement is the smarter moving cursor back. The limit to moving the cursor back is height of the stdout. If we need to update things that are no longer accessible, we add 3 new lines with the text 'Display Updated' to show that new changes were added. We then redraw the changed lines. This method maintains the efficiency of cacheing lines, and now writing to stdout more than neccessary.

Another improvement is on the close() method. The improvement also resets the cursor x.